+++ src/vec/operations.h
@@ -2,6 +2,7 @@
 #define VEC_OPERATIONS_H
 
 #include <cstdint>
+#include <stdexcept>
 #include <vector>
 
 namespace operations {
@@ -162,207 +163,48 @@ class Base256 {
         data = std::move(result);
     }
 
-    void add(const Base256& rhs) {
-        add(rhs.data);
-    }
-
-    Base256& operator+=(const Base256& rhs) {
-        add(rhs);
-        return *this;
-    }
-
-    void sub(const Base256& rhs) {
-        sub(rhs.data);
-    }
-
-    Base256& operator-=(const Base256& rhs) {
-        sub(rhs);
-        return *this;
-    }
-
     // The return value can only be positive, if it would be negative, 0 is returned
     void sub(const std::vector<std::uint8_t> &b) noexcept {
         // Prevent from ending the subtraction before going over the hole subtractor and stop if the
         // result can only be negative
         if (getStartBitIndex(b) > getStartBitIndex(data)) throw std::invalid_argument("Invalid base256 index");
 
-        std::vector<std::uint8_t> result;
-
-        // Handle an underflow when subtracting
-        bool borrow = false;
-
-        for (int i = 0; i < data.size(); i++) {
-            std::int32_t subtract;
-            // Check for the end of the subtractor
-            if (i >= b.size()) {
-                subtract = borrow;
-            } else {
-                subtract = borrow + b[i];
-            }
-            borrow = false;
-
-            if (data[i] >= subtract) {
-                // If the current number is as least as big as subtract
-                std::uint8_t number = data[i] - subtract;
-                result.push_back(number);
-            } else {
-                // Borrow from the next number
-                subtract -= 256;
-                borrow = true;
-
-                // Here subtract can only be 0 or negative
-                std::uint8_t number = data[i] - subtract;
-                result.push_back(number);
-            }
-        }
+        std::vector<std::uint8_t> result = sub(data, b);
 
         data = std::move(result);
     }
 
-    [[nodiscard]] std::vector<std::uint8_t> convertToVector(std::uint64_t number) noexcept {
-        std::vector<std::uint8_t> result;
+    [[nodiscard]] std::vector<std::uint8_t> convertToVector(std::uint64_t number) noexcept;
 
-        // Loop until the whole number is zero
-        while (number) {
-            // We only care for the lsb
-            result.push_back(static_cast<std::uint8_t>(number & 0xFF));
+    // The return value can only be positive, if it would be negative, 0 is returned
+    [[nodiscard]] std::vector<std::uint8_t> sub(const std::vector<std::uint8_t> &a,
+                                                const std::vector<std::uint8_t> &b) noexcept;
 
-            // Shift the number by 8 to the right
-            number >>= 8;
-        }
+    void mul(const std::vector<std::uint8_t> &b) noexcept;
+    void div(const std::vector<std::uint8_t> &divisor,
+        std::vector<std::uint8_t> *remaining = nullptr) noexcept;
 
-        return result;
-    }
+    [[nodiscard]] std::vector<std::uint8_t> pow(const std::vector<std::uint8_t> &a,
+                                                const std::uint64_t &pow) noexcept;
 
-    void mul(const Base256& rhs) {
-        mul(rhs.data);
-    }
-
-    Base256& operator*=(const Base256& rhs) {
-        mul(rhs);
+    Base256& operator+=(const Base256& rhs) {
+        add(rhs.data);
         return *this;
     }
 
-    void mul(const std::vector<std::uint8_t> &b) noexcept {
-        // Time complexity O(aSize * bSize)
-        const std::uint64_t aSize = data.size();
-        const std::uint64_t bSize = b.size();
-
-        // Initialize the result vector with zeros, with the size of aSize + bSize
-        std::vector<std::uint8_t> result(aSize + bSize, 0);
-
-        for (std::uint64_t i = 0; i < aSize; i++) {
-            // Set up the carry value for each iteration
-            std::uint16_t carry = 0;
-
-            for (uint64_t x = 0; x < bSize; x++) {
-                // Calculate the product by adding up the previous result, the carry, and the new
-                // product
-                std::uint16_t product = result[i + x] + carry + (data[i] * b[x]);
-
-                // Calculate the carry which is the overflow beyond 255
-                carry = product >> 8;
-
-                // Store the least significant byte (lsb) of the product in the result
-                result[i + x] = static_cast<std::uint8_t>(product & 0xFF);
-            }
-
-            // If there is a carry, append it to the result
-            // But with the offset of bSize because of the previous inner loop
-            result[i + bSize] += static_cast<std::uint8_t>(carry);
-        }
-
-        data = std::move(result);
+    Base256& operator-=(const Base256& rhs) {
+        sub(rhs.data);
+        return *this;
     }
 
-    [[nodiscard]] std::vector<std::uint8_t> div(
-        const std::vector<std::uint8_t> dividend, const std::vector<std::uint8_t> &divisor,
-        std::vector<std::uint8_t> *remaining = nullptr) noexcept {
-        std::vector<std::uint8_t> quotient;
-        std::uint8_t quotientBuffer = 0;
-        std::uint16_t quotientBitIndex = 0;
-
-        // The index of the last bit in the dividendMask inside dividend
-        std::int64_t dividendIndex = getStartBitIndex(dividend);
-        // This copy's the most significant bit of the dividend
-        std::vector<std::uint8_t> dividendMask = addBitFromNumber({0}, dividend, dividendIndex--);
-
-        while (dividendIndex >= -1) {
-            if (quotientBitIndex > 7) {
-                // The quotient is stored from the most significant byte to the least significant
-                // byte, in contrast to all the other vector based numbers.
-                // Which is later reversed, at the end of the function.
-                quotient.push_back(quotientBuffer);
-                quotientBuffer = 0;
-                quotientBitIndex = 0;
-            }
-
-            if (isEqual(dividendMask, divisor) || isBigger(dividendMask, divisor)) {
-                // Stop the loop if the dividend is smaller than the divisor, because fractional
-                // digits are not supported
-                if (dividendIndex < 0) {
-                    quotientBuffer <<= 1;
-
-                    // If remaining pointer is passed, set the remaining value
-                    if (remaining != nullptr) *remaining = dividendMask;
-                    break;
-                }
-
-                // Shift the dividend and set the new bit as high
-                quotientBuffer <<= 1;
-                quotientBuffer++;
-                quotientBitIndex++;
-
-                // TODO
-                //dividendMask = sub(dividendMask, divisor);
-                dividendMask = addBitFromNumber(dividendMask, dividend, dividendIndex--);
-            } else {
-                // Stop the loop if the dividend is smaller than the divisor, because fractional
-                // digits are not supported
-                if (dividendIndex < 0) {
-                    quotientBuffer <<= 1;
-
-                    // If remaining pointer is passed, set the remaining value
-                    if (remaining != nullptr) *remaining = dividendMask;
-                    break;
-                }
-
-                // if the divisor is bigger than the dividend, we need to shift the dividend and set
-                // the new bit as low
-                quotientBuffer <<= 1;
-                quotientBitIndex++;
-
-                // Stop the loop if the dividend is smaller than the divisor, because fractional
-                // digits are not supported
-                if (dividendIndex < 0) break;
-
-                // Because dividendMask is smaller than divisor, we add the next bit from the
-                // dividend
-                dividendMask = addBitFromNumber(dividendMask, dividend, dividendIndex--);
-            }
-        }
-
-        if (quotientBuffer != 0) quotient.push_back(quotientBuffer);
-
-        // Reverse the quotient
-        std::reverse(quotient.begin(), quotient.end());
-
-        return quotient;
+    Base256& operator*=(const Base256& rhs) {
+        mul(rhs.data);
+        return *this;
     }
 
-    [[nodiscard]] std::vector<std::uint8_t> pow(const std::vector<std::uint8_t> &a,
-                                                const std::uint64_t &pow) noexcept {
-        // Copy the value from a into result while keeping a constant
-        std::vector<std::uint8_t> result;
-        std::copy(a.begin(), a.end(), std::back_inserter(result));
-
-        // Start the loop at 1, because the first number is already assigned to result
-        for (std::uint32_t i = 1; i < pow; i++) {
-            // TODO
-            //result = mul(result, a);
-        }
-
-        return result;
+    Base256& operator/=(const Base256& rhs) {
+        div(rhs.data);
+        return *this;
     }
 
 private:
